feat: consolidate and reorganize constants into dedicated files for improved structure and maintainability
